% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/physet-class.R
\docType{methods}
\name{access}
\alias{$,physet-method}
\alias{Enames}
\alias{Enames,physet-method}
\alias{Nedge}
\alias{Nedge,physet-method}
\alias{Nsample}
\alias{Nsample,physet-method}
\alias{Ntaxa}
\alias{Ntaxa,physet-method}
\alias{Snames}
\alias{Snames,physet-method}
\alias{Tnames}
\alias{Tnames,physet-method}
\alias{access}
\alias{access,physet-method}
\alias{edge_com}
\alias{edge_com,physet-method}
\alias{edge_len}
\alias{edge_len,physet-method}
\alias{edge_mat}
\alias{edge_mat,physet-method}
\alias{otu_table}
\alias{otu_table,physet-method}
\alias{phy_tree}
\alias{phy_tree,physet-method}
\alias{sample_data}
\alias{sample_data,physet-method}
\alias{seqdep}
\alias{seqdep,physet-method}
\alias{tax_table}
\alias{tax_table,physet-method}
\title{Access slots or attributes in physet object.}
\arguments{
\item{x}{A physet object}

\item{slot_name}{A character indicating the slot name: \code{otu_table},
\code{sample_data}, \code{tax_table}, \code{seqdep}, \code{phy_tree},
\code{edge_com}, \code{edge_len}, \code{edge_mat}. The function will return
\code{NULL} if \code{slot_name} does not exist. \code{Nsample}, \code{Ntaxa},
\code{Nedge} are allowed to access the count of samples, taxa and utilized
edges. \code{Snames}, \code{Tnames} and \code{Enames} can return sample
names, taxa names and utilized tree edge names in physet object.}

\item{name}{A character specifying variables. \code{name} can be any value
in \code{names(sample_data(x))} and \code{names(tax_table(x))}. If
\code{name} exists in both \code{sample_data} and \code{tax_table}, this will
return a vector contains both columns with this \code{name}.}
}
\value{
Corresponding object or information for \code{name} of \code{x}.
}
\description{
Access slots in physet object by slot name. Or get attributes in
sample_data and tax_table like name and size.
}
\section{Usage}{

access(x, slot_name) \cr
slot_name(x) \cr
x$name \cr
}
\examples{
data(oral)
access(oral, "sample_data")
seqdep(oral)
oral$Group
Ntaxa(oral)
Snames(oral)
}
\seealso{
\code{\link{assign}}, \code{\link{extract}}
}

