% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal-methods.R
\name{.alphaEst}
\alias{.alphaEst}
\title{Estimate alpha diversity, richness and eveness}
\usage{
.alphaEst(x, split = TRUE, ...)
}
\arguments{
\item{x}{A \code{matrix} or \code{physet} object.}

\item{split}{treat samples separately (\code{TRUE}) or as one community.}

\item{...}{not used}
}
\description{
A function used to estimate richness for a given object.
}
\examples{
data(oral)
.alphaEst(oral)
}
\seealso{
\code{\link[vegan]{diversity}}, \code{\link[phyloseq]{estimate_richness}}.
}

