% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/physet-class.R
\docType{methods}
\name{assign}
\alias{$<-,physet-method}
\alias{Snames<-}
\alias{Snames<-,physet-method}
\alias{Tnames<-}
\alias{Tnames<-,physet-method}
\alias{assign}
\alias{otu_table<-}
\alias{otu_table<-,physet-method}
\alias{phy_tree<-}
\alias{phy_tree<-,physet,phyloOrNULL-method}
\alias{sample_data<-}
\alias{sample_data<-,physet-method}
\alias{seqdep<-}
\alias{seqdep<-,physet,numeric-method}
\alias{tax_table<-}
\alias{tax_table<-,physet-method}
\title{Update or slots or attributes in physet object.}
\arguments{
\item{x}{A physet object.}

\item{value}{A new object with supported class for \code{name}.}

\item{name}{Available slot name are: \code{otu_table}, \code{sample_data},
\code{tax_table}, \code{phy_tree}, \code{seqdep}. Any attributes in
\code{names} for \code{sample_data(x)} and \code{tax_table(x)} can be used.
If \code{name} is a new attributes for \code{x} object, specify the slot
by using suffix \dQuote{._AddToS_} for \code{sample_data} and suffix
\dQuote{._AddToT_} for \code{tax_table} (the suffix will be removed).}
}
\value{
A new \code{\link{physet-class}} object with updated information.
}
\description{
Assign slots or attributes with new information or object. The
function will automatically update related slots based on changes.
}
\details{
The assign methods for updating whole slots actually generate a
new \code{\link{physet}} object by passing the original object to the
\code{\link{physet-constructor}}. While assign methods for adding variables
or change names will keep the original one. \cr
New \code{sample_data} and \code{tax_table} are actually added to the same
slot in the original \code{physet}object. Only those conflict attributes
are replaced. Use \dQuote{._AddToS_} and \dQuote{._AddToT_} suffix to add
extra attributes. \cr
Assign new values to \code{edge_com}, \code{edge_len} and \code{edge_mat}
manually are not recommended as they are linked with \code{otu_table} and
\code{py_tree}. Provide a new \code{otu_table} or \code{phy_tree} or both
automatically trigger the \code{\link{physet-constructor}} to re-build
these information securely.
}
\section{Usage}{

name(x) <- value
x$name <- value
}
\examples{
data(oral)
data(oral_raw)
otu_table(oral) <- oral_raw$rawdata[, 1:4]
oral$newtaxaID._AddToT_ <- paste("ID", c(1:Ntaxa(oral)), sep="")
}
\seealso{
\code{\link{access}}, \code{\link{extract}},
\code{\link{physet-constructor}}
}

