% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrapper-methods.R
\name{baseinfo_analysis}
\alias{baseinfo_analysis}
\title{Abundance and Cont table analysis wrapper}
\usage{
baseinfo_analysis(x, formula, opts, ...)
}
\arguments{
\item{x}{A \code{\link{physet-class}} objects.}

\item{formula}{A formula used for constrain analysis.}

\item{opts}{The ploting options generated by \code{\link{phyplotOptions}}.}

\item{...}{other parameters. see Details.}
}
\value{
A large list with the following structure:
}
\description{
Basic taxa abundance and alpha diversity analysis.
}
\note{
support all vegdist and unifrac \cr
Mental like Test \cr
Intra-group vs. Inter-group distance average \cr
PCoA, NMDS, 2d + 3d \cr
}
\examples{
data(oral)
opts <- phyplotOptions(colorVar = list(sample = "Group", taxa = "Domain"),
                       shapeVar = list(sample = "Periodontitis"),
                       fillVar = list(sample = "Group", taxa = "Domain"))
res <- baseinfo_analysis(oral, . ~ Group, opts, rarefy = TRUE, permu = 100)
res$tax.boxplot$jitter
res$tax.barplot$outlier
res$diver.plot$tax.alphaindex$boxplot
res$diver.plot$tax.rarefy$group
}

