% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal-methods.R
\name{.phycolors}
\alias{.phycolors}
\alias{.physhapes}
\title{Generate colors and shapes for a given vector}
\usage{
.phycolors(x, value, ...)

.physhapes(x, value, ...)
}
\arguments{
\item{x}{A list of 2 vectors or factors, each unique element will be assigned
with a unique color/shape.}

\item{value}{The real colors used for plotting. The function will generate
a sequence of color/shape if \code{value} is missing.}

\item{...}{other parameters, not used}
}
\description{
Generate colors and shapes for a given vector
}
\keyword{internal}

