% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal-methods.R
\docType{methods}
\name{.cumNode}
\alias{.cumNode}
\alias{.cumNode,phylo-method}
\alias{.cumNode.phylo}
\title{Generate Tree Node abundance matrix}
\usage{
.cumNode(tree, ...)

\S4method{.cumNode}{phylo}(tree, ...)
}
\arguments{
\item{tree}{A \code{\link[ape]{phylo}} object, or a \code{phy_tree} slot.}

\item{...}{unused parameters.}
}
\value{
A sparse boolean \code{Matrix} stored taxa composition for nodes
and leafs.
}
\description{
A internal function used to generate edge structure of tree and
edge abundance matrix for further analysis like unifrac.
}
\details{
The function is automatically called in
\code{\link{physet-constructor}} when \code{phy_tree} is provided.
Fast-Unifrac algorithm is used for parallel computation.
}
\keyword{internal}

