% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/physet-class.R
\docType{methods}
\name{show,physet-method}
\alias{show,physet-method}
\title{Show method for physet object}
\usage{
\S4method{show}{physet}(object)
}
\arguments{
\item{object}{A physet object}
}
\description{
The function to show or print a physet object.
}
\examples{
data(oral)
oral
print(oral)
}
\seealso{
\code{\link{summary}}
}
\keyword{internal}

