% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrapper-methods.R
\name{enterotype_analysis}
\alias{enterotype_analysis}
\title{Enterotype analysis for objects}
\usage{
enterotype_analysis(x, method = "JSD", ...)
}
\arguments{
\item{x}{A matrix, or \code{\link{physet-class}} objects.}

\item{method}{The distance method used for pam clustering.}

\item{...}{other parameters passed to clustering \code{\link[cluster]{pam}}
and distance method \code{\link{phydist}}.}
}
\value{
A large list with the following structure:
}
\description{
This is the enterotype analysis wrapper based on the code in
enterotype official website.
}
\examples{
data(oral)
res <- enterotype_analysis(oral)
plot(res$nclusters, type = "h", xlab = "k clusters", ylab = "CH index")
res$data.cluster
res$distance
}
\references{
enterotype official website
}

