% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/physet-class.R
\docType{methods}
\name{[,physet,ANY,ANY,ANY-method}
\alias{[,physet,ANY,ANY,ANY-method}
\alias{extract}
\title{Extract Parts of an physet object}
\usage{
\S4method{[}{physet,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A physet object.}

\item{i,j,...}{Indices specifying elements to extract: samples, taxa and
edges. The function works as same as extractor in \code{\link{Extract}},
but it always set \code{drop = FALSE} to maintain slot's structure.}

\item{drop}{This variable is fixed as \code{FALSE}.}
}
\value{
A subset of original \code{\link{physet-class}} object.
}
\description{
Extract from a physet object
}
\examples{
library(Matrix)
data(oral)
oral[1:10, 2:5]
oral[1:5, grep("^[AP]", oral$Phylum),
     colMeans(edge_mat(oral)/seqdep(oral)) > 0.3]
}
\seealso{
\code{\link{access}}, \code{\link{assign}}
}

