% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrapper-methods.R
\name{feature_selection}
\alias{feature_selection}
\title{Supervised Learning Feature Selection}
\usage{
feature_selection(data, group, methods = c("StaTest", "RF", "ENet", "ISA",
  "SVMRFE"), ...)
}
\arguments{
\item{data}{A matrix, or \code{\link{physet-class}} objects.}

\item{group}{The grouping variable or a factor.}

\item{methods}{The list of selection methods.}

\item{...}{other parameters passed to feature selection methods. Probably a
kernel function with \code{kernel = "linear"}}
}
\value{
A large list with the following structure:
}
\description{
This is the wrapper for several feature selection methods.
}
\details{
Statistical Test: T-test, Fisher exact test,
Np T-test, Multi-group comparison \cr
Random Forest Selection: Standard entropy method, Boruta selection \cr
ENET Selection: ENET Logistic Regression, DrSVM, HHSVM \cr
SVM-RFE: 2-group, multi-group \cr
Indicator Species: labdsv \cr
}

