% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbr.R
\docType{package}
\name{microbr-package}
\alias{microbr-package}
\title{Dong Lab Microbiome Analysis Pipeline.}
\description{
The microbr package is designed is a general purpose wrapper and pipeline
for lots of routine microbiome and metagenomics analysis. The pipeline takes
meta data, abundance table, taxonomy information and phylogenetic tree from
upstream sequencing analysis as input and simplify downstream data exploring.
\tabular{ll}{
Diversity analysis: \tab alpha/beta-diversity, richness, eveness \cr
Univariate analysis: \tab (non)-parametric, multi-group, p-value correction \cr
Multivariate analysis: \tab ordination analysis, distance analysis \cr
Enterotype analysis: \tab enterotype clustering \cr
Machine learning: \tab classification, regression, feature selection, clustering \cr
}
}
\details{
\tabular{ll}{
Package: \tab microbr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-26\cr
License: \tab UNT\cr
}
}
\section{Important}{

This package is used in Dong's Lab only or for personal learning purpose.
Similar packages (like \code{\link[phyloseq]{phyloseq}}) have been developed and well
maintained in Github and Bioconductor. Some codes are copied from or highly
similar to other contributions. Please \emph{CONTACT the authors} for
permission if codes will be published or be used for other purpose.
}
\examples{
## ~~ simple examples of the most important functions ~~
}
\author{
Developer: Ruichen Rong <\email{imprompturong@gmail.com}> \cr
Maintainer: Kashi <\email{}>, Huaiying <\email{}>, Yuqing<\email{}>
}
\references{
\url{Too Many}
}
\seealso{
\code{\link[phyloseq]{phyloseq-package}}
}
\keyword{package}

