% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbr.R
\name{oral}
\alias{oral}
\title{Oral Microbiome Dataset - physet object}
\format{A physet object with otu_table, sample_data, tax_table, phy_tree.}
\source{
\url{http://aaa/}
}
\description{
The physet object constructed from \code{data(oral_raw)}. It is
a sample physet object for testing purpose. See \code{\link{physet-class}}
and \code{\link{physet}} for details.
}

