% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbr.R
\name{oral_raw}
\alias{oral_raw}
\title{Oral Microbiome Dataset - raw input}
\format{A list with 4 slots for oral microbiome.}
\source{
\url{http://aaa/}
}
\description{
A microbiome dataset containing the phylum count table, sample
information, phyla information and phylogenetic tree.
The variables are as follows:

\itemize{
  \item rawdata. The phylum count table for Oral data.
  \item metainfo. The meta information for samples.
  \item taxonomy. The taxa information for phyla.
  \item rawtree. The phylogenetic tree on phylum level.
}
}

