% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrapper-methods.R
\name{ordinate_analysis}
\alias{ordinate_analysis}
\title{Distance and Ordination analysis wrapper}
\usage{
ordinate_analysis(x, formula, opts, methods, ...)
}
\arguments{
\item{x}{A \code{\link{physet-class}} objects.}

\item{formula}{A formula used for constrain analysis.}

\item{opts}{The ploting options generated by \code{\link{phyplotOptions}}.}

\item{methods}{A set of method name passed to \code{\link{phydist}}.}

\item{...}{other parameters. see Details.}
}
\value{
A large list with the following structure:
}
\description{
Distance and Ordination analysis wrapper
}
\note{
support all vegdist and unifrac \cr
Mental like Test \cr
Intra-group vs. Inter-group distance average \cr
PCoA, NMDS, 2d + 3d \cr
}
\examples{
data(oral)
opts <- phyplotOptions(colorVar = list(sample = "Group", taxa = "Domain"),
                       shapeVar = list(sample = "Periodontitis"),
                       fillVar = list(sample = "Group", taxa = "Domain"))
res <- ordinate_analysis(oral, . ~ Group, opts, "unifrac.w.un")
res$biplot
res$splitplot
}

