% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal-methods.R
\name{.pcorrect}
\alias{.pcorrect}
\title{Multiple Test Correction}
\usage{
.pcorrect(p, ...)
}
\arguments{
\item{p}{probability matrix}

\item{...}{other parameters passed to \code{\link{p.adjust}}}
}
\description{
Multiple Test Correction
}
\keyword{internal}

