% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ordinate-methods.R
\docType{methods}
\name{phydist}
\alias{phydist}
\alias{phydist,physet-method}
\alias{phydist.data.frame}
\alias{phydist.default}
\alias{phydist.physet}
\title{Distance and Dissimilarity}
\usage{
phydist(x, method = "bray", diag = FALSE, upper = FALSE, ...)

\S4method{phydist}{physet}(x, method = "bray", diag = FALSE,
  upper = FALSE, ...)
}
\arguments{
\item{x}{A \code{matrix} or any object that inherits this method: like
\code{\link{physet-class}}}

\item{method}{the distance measure to be used. Default method is "bray".
Supported distances include: unifrac.uw, unifrac.w.n, unifrac.w.un, JSD and
all methods registered in \code{\link[vegan]{vegdist}}.}

\item{diag}{logical value indicating whether the diagonal of the distance
matrix should be printed.}

\item{upper}{logical value indicating whether the upper triangle of the
distance matrix should be printed.}

\item{...}{all other parameters passed to \code{vegdist} or \code{dist}.}
}
\value{
A distance matrix of class \code{\link{dist}}.
}
\description{
Calculate distance or dissimilarity for community data.
}
\examples{
data(oral)
dist <- phydist(oral, method = "unifrac.w.un")
}
\seealso{
\code{\link[vegan]{vegdist}}, \code{\link[phyloseq]{distance}}
}

