% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{phyplotOptions}
\alias{phyplotOptions}
\title{Physet Ploting Options}
\usage{
phyplotOptions(...)
}
\arguments{
\item{...}{The sample and taxa information if available. \cr
Current supported plotting options: \cr
\itemize{
\item{opts}{ Update current \code{opts} with new parameters. }
\item{info}{ Build up \code{opts} with new sample and taxa info. }
\item{color}{ The colors used for plotting. }
\item{colorVar}{ The variable used to specify colors. }
\item{fillVar}{ The variable used to specify fill colors for shapes. }
\item{shape}{ The shapes used for plotting. }
\item{shapeVar}{ The variable used to specify shapes. }
\item{labelVar}{ The variable used to specify labels. }
\item{axes.2d}{ Axes ID as \dQuote{x} and \dQuote{y} for 2d plot. }
\item{axes.3d}{ Axes ID as for 3d plot. }
\item{keepOnly}{ A list with attributes \code{sample} and \code{taxa}. This
parameter decides which sample and taxa are showned in figures. }
\item{heatmapscale}{ The scale used to generate heatmap color. }
\item{heatmaptrans}{ The pre-treatment method to data in heatmap. }
\item{outputName}{ The prefix name used for output tables and figures. }
\item{...}{ Unused parameters for now. Possibly added to make the function
compatiblewith old source code for igraph and 3d plotting. }
}}
}
\value{
A list includes all necessary plotting parameters for plot method
in \code{microbr}.
}
\description{
This function is used to generate ploting options for objects.
}
\examples{
data(oral)
colorVar <- list(sample = "Group", taxa = "Domain")
shapeVar <- list(sample = "Periodontitis")
fillVar <- colorVar
opts <- phyplotOptions(info = oral, colorVar = colorVar, shapeVar = shapeVar, fillVar = fillVar)
ord <- ordinate(oral, "NMDS", method = "unifrac.w.un")
res <- plot(ord, opts)
res$biplot
res$splitplot
}
\seealso{
\code{link{plot.ordination}}
}

