% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/physet-class.R
\docType{class}
\name{physet-class}
\alias{physet}
\alias{physet-class}
\title{The S4 class for microbiome analysis}
\description{
The \code{physet-class} is used to store most of the microbiome
components in a single instance. The physet class allows easy and flexible
manipulation on sample/taxa and it can facilitate downstream analyses.

The \code{\link{physet-constructor}} ensures that different data components
are compatible. Initially slots are set to \code{NULL}. The 4 primary slots
can be updated with corresponding assignment methods while additional slots
can be updated with phylogenetic tree. Also \code{\link{extract}} method can
be used to exclude unwanted sample/taxa from downstream analyses.

The \code{physet-class} contains 4 primary slots. They are the fundemental
components for most of the downstream analyses.
Primary slots:
}
\section{Slots}{

\describe{
\item{\code{otu_table}}{A [\code{Nsample * Ntaxa}] \code{\link[Matrix]{Matrix}} for
count table or abundance table. The otu_table will keep raw count or
abundance and it will be normalized by \code{seqdep} automatically
for most statistical analysis. The otu_table could also be randomly
subsampled with \code{\link{.subsampling}} (not recommanded, see
reference).
New otu_table can be assigned with \code{\link{otu_table<-}}.}

\item{\code{sample_data}}{A \code{\link{data.frame}} for meta data information. This
slot is matched accroding to \code{rownames(otu_table)}. So samples
only occured in \code{otu_table} will have \code{NA} attributes while
samples not occured in \code{otu_table} will be eliminated. One can
use \code{\link{extract}} method to manipulate unwanted samples. New
\code{sample_data} can be assigned by \code{\link{sample_data<-}}.}

\item{\code{tax_table}}{A \code{\link{data.frame}} for taxonomy information. This
slot is matched accroding to \code{colnames(otu_table)}. So taxa only
occured in \code{otu_table} will have \code{NA} attributes while taxa
not occured in \code{otu_table} will be eliminated. One can use
\code{\link{extract}} method to eliminate unwanted taxa. New taxonomy
information can be assigned with \code{\link{tax_table<-}}.}

\item{\code{seqdep}}{Sequencing depth of each sample for normalization. In order to
       keep analysis under same scope, this slot will not be changed when
       you eliminate unwanted taxa. So use \code{\link{seqdep<-}} to modify
       it if you want to re-normalize \code{otu_table}, \code{edge_mat}.

When phylogenetic tree is provided, 4 additional slots are added for
phylogeny based analyses like unifrac and dpcoa.
Additional slots:}

\item{\code{phy_tree}}{A \code{\link[ape]{phylo}}-class from the ape package for
a rooted phylogenetic tree. Assign value to \code{phy_tree} will
not modify the 6 primary slot. This approach guarantee that tree
is kept without trimming and it will not influence non-phylogeny
analysis. \code{physet} will not modify \code{phy_tree}, please
check \pkg{\link[ape]{ape}} package for tree manipulation.}

\item{\code{edge_len}}{A named vector stored the branch length for each edge in
\code{phy_tree}.}

\item{\code{edge_com}}{A boolean \code{\link[Matrix]{Matrix}} stored hierarchcal
structure of input tree. It can simplify some analysis and it is
useful to generate supporting files for iTOL plot.}

\item{\code{edge_mat}}{A [\code{Nsample * Nedge}] \code{\link{Matrix}} records the
       total taxa abundance under each node for each sample. Any change in
       \code{otu_table} and \code{phy_tree} will modify the value. The
       matrix is used to speed up unifrac, dpcoa diversity calculation and
       is used for phylogeny based feature selection.}
}}
\examples{
getSlots("physet")
slotNames("physet")
}
\seealso{
The constructor \code{\link{physet-constructor}},
 accessors \code{\link{otu_table}}, \code{\link{sample_data}},
 \code{\link{tax_table}}, \code{\link{phy_tree}}.
}

