% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/physet-class.R
\docType{methods}
\name{initialize,physet-method}
\alias{initialize,physet-method}
\alias{physet-constructor}
\title{Build a physet-class from its components}
\arguments{
\item{.Object}{The physet object to create.}

\item{...}{Components to be passed to the constructor, 4 major components,
phylogenetic tree and another \code{physet} class are allowed as input.
The constructor has no order for inputs. Specify input parameter with its
slot name in physet: "otu_table=yourdata, phy_tree=yourtree".
 \itemize{
   \item{physet}{ a \code{\link{physet}} object. Further arguments will
         be used to update this object. Use \code{physet = }. }
   \item{otu_table}{ a \code{matrix}, \code{Matrix} or \code{data.frame}
         for taxa table. Use \code{otu_table = }. }
   \item{sample_data}{ a \code{\link{data.frame}} for sample information.
         All \code{factor} will be flatterned as \code{atomic} and will be
         transfered back when needed. This will avoid null level problem.
         If \code{physet} is provided, the \code{sample_data} will expand
         not replace the original \code{sample_data} in \code{physet}. Only
         conflict attributes will be overwritten. Use \code{sample_data = }. }
   \item{tax_table}{ a \code{\link{data.frame}} for taxonomy information.
         All \code{factor} will be flatterned as \code{atomic} and will be
         transfered back when needed. This will avoid null level problem.
         If \code{physet} is provided, the \code{sample_data} will expand
         not replace the original \code{sample_data} in \code{physet}. Only
         conflict attributes will be overwritten. Use \code{tax_table = }. }
   \item{seqdep}{ a \code{numeric} value or a \code{numeric vector} with
         length \code{Nsample}. By default, the constructor will use
         \code{rowSums(otu_table)} as \code{seqdep} if not specified. Use
         \code{seqdep = }. }
 }}
}
\value{
A physet object is returned. All slots are \code{NULL} if nothing is
 passed. Otherwise returned object will at least contains 4 primary slots.
}
\description{
The \code{physet-constructor} method is the initialization method to create
a valid \code{\link{physet-class}}. This function firstly manage 4 major
components defined in \code{\link{physet-class}}. Then if a phylogenetic
tree is given, the internal function \code{\link{.cumNode}} is called to set
up 3 additional components. Unprovided slots will be kept as \code{NULL} and
can be assigned later. Also any physet instance can be re-build with the
constructor with new slots.
}
\section{Usage}{

physet(...) \cr
## S4 method for signature 'physet' \cr
initialize(.Object, ...) \cr
}
\examples{
require(foreach)
data(oral_raw)
oral <- physet(otu_table=oral_raw$rawdata, sample_data=oral_raw$metainfo,
               tax_table=oral_raw$taxonomy, phy_tree=oral_raw$rawtree)
oral
slotNames(oral)
}
\seealso{
\code{\link{access}}, \code{\link{assign}}, \code{\link{extract}},
}

