% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.ordination}
\alias{plot.ordination}
\title{Plotting method for Ordination Object}
\usage{
\method{plot}{ordination}(x, opts, info, ...)
}
\arguments{
\item{x}{A ordination object generated by \code{\link{ordinate}}.}

\item{opts}{The plotting options generated by \code{\link{phyplotOptions}}.}

\item{info}{A object contains sample and taxa information. This parameter is
used to update \code{opts} with . \code{info} be a list
object with name \dQuote{sample} and \dQuote{taxa}. Or any object inherits
\code{\link{physet}} class.}

\item{...}{Other parameters passed to ordination and phydist function.}
}
\value{
A list of objects including: biplot, splitplot(sample and taxa in 2
different panel with same scale), eigen plot, heatmap and plot data frame.
(Note: 3D plot are outputed inside the function because it cannot be saved
as a valid object.)
}
\description{
The function used to plot ordination class.
}
\examples{
data(oral)
colorVar <- list(sample = "Group", taxa = "Domain")
shapeVar <- list(sample = "Periodontitis")
fillVar <- colorVar
opts <- phyplotOptions(colorVar = colorVar, shapeVar = shapeVar, fillVar = fillVar)
ord <- ordinate(oral, "CAP", . ~ Group, method = "unifrac.w.un")
res <- plot(ord, opts, oral)
res$biplot
res$splitplot
res$eigenplot
res$plotDF
}
\seealso{
\code{\link[phyloseq]{plot_ordination}},
\code{\link[phyloseq]{plot_phyloseq}}
}

