% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.rarefy}
\alias{plot.rarefy}
\title{Plotting method for Rarefaction Object}
\usage{
\method{plot}{rarefy}(x, opts, info, ...)
}
\arguments{
\item{x}{A ordination object generated by \code{\link{ordinate}}.}

\item{opts}{The plotting options generated by \code{\link{phyplotOptions}}.}

\item{info}{A object conains sample and taxa information. This can be a list
object with name \dQuote{sample} and \dQuote{taxa}. Or any object inherits
\code{\link{physet}} class.}

\item{...}{Other parameters passed to ordination and phydist function.}
}
\value{
A list of objects including: biplot, splitplot(sample and taxa in 2
different panel with same scale), eigen plot, heatmap and plot data frame.
}
\description{
The function used to plot rarefaction curves.
}
\examples{
data(oral)
colorVar <- list(sample = "Group", taxa = "Domain")
shapeVar <- list(sample = "Periodontitis")
fillVar <- colorVar
opts <- phyplotOptions(colorVar = colorVar, shapeVar = shapeVar, fillVar = fillVar)
raf <- rarefaction(oral, . ~ Group, rarefy = TRUE, permutation = 100)
res <- plot(raf, opts, oral)
res$tax.richness$boxplot
res$tax.richness$barplot
res$tax.alphaindex$boxplot
res$tax.alphaindex$barplot
res$tax.rarefy$group
res$tax.rarefy$sample
}
\seealso{
\code{\link[phyloseq]{plot_richness}},
\code{\link[phyloseq]{plot_phyloseq}}
}

