% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/support-methods.R
\name{rarefaction}
\alias{rarefaction}
\title{Rarefaction Analysis}
\usage{
rarefaction(x, formula, data = sample_data(x), rarefy = FALSE,
  permu = 100, ...)
}
\arguments{
\item{x}{A \code{\link{physet-class}} objects.}

\item{formula}{A formula used for constrain analysis.}

\item{data}{The meta information contains RHS of formula.}

\item{rarefy}{Whether using bootstrap random sampling to generate plotdata
for rarefaction curves.}

\item{permu}{Times for bootstrap, only used if \code{rarefy = TRUE}.}

\item{...}{other parameters. not used.}
}
\value{
A large list with the following structure:
}
\description{
Rarefaction Analysis
}
\note{
support shannon, simpson, pielou, chao1, ace \cr
Diversity statistical test \cr
}
\examples{
data(oral)
rarefaction(oral, . ~ Group, rarefy = TRUE, permutation = 100)
}

