% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/support-methods.R
\name{phyTtest}
\alias{phyCtest}
\alias{phyMtest}
\alias{phyTtest}
\title{Statistical Analysis for physet object}
\usage{
phyTtest(x, formula, data = sample_data(x), ...)

phyCtest(x, formula, data = sample_data(x), ...)

phyMtest(x, formula, data = sample_data(x), ...)
}
\arguments{
\item{x}{A \code{matrix} or a \code{\link{physet-class}} objects.}

\item{formula}{A formula used for constrain analysis.}

\item{data}{The data frame contains the grouping variable.}

\item{...}{other parameters. not used.}
}
\value{
A large list with the following structure:
}
\description{
These methods are used to calculate univariate statistical
analysis. Including T-test, Wilcoxon-test, Fisher-exact-test, Macnemar-test,
Anova, Kruskal test, Tukey HSD and Mann's NP multiple group comparison.
}
\examples{
data(oral)
phyTtest(oral, . ~ Group)
phyCtest(oral, . ~ Group)
phyMtest(oral, . ~ Group + Sex)
}

