% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal-methods.R
\docType{methods}
\name{.subsampling}
\alias{.subsampling}
\alias{.subsampling,physet-method}
\alias{.subsampling.character}
\alias{.subsampling.double}
\alias{.subsampling.integer}
\alias{.subsampling.matrix}
\alias{.subsampling.physet}
\title{Random Sampling and Shuffling}
\usage{
.subsampling(x, ...)

.subsampling.integer(x, size, rep, ...)

.subsampling.double(x, size, rep, ...)

.subsampling.character(x, size, rep, ...)

.subsampling.matrix(x, size, rep, ...)

\S4method{.subsampling}{physet}(x, size, rep, ...)
}
\arguments{
\item{x}{A \code{numeric}, \code{character} \code{vector} or \code{matrix}.
It can also be any object that inherits this method, like
\code{\link{physet-class}}.}

\item{...}{all other parameters. Not available for now.}

\item{size}{The sampling pool size, sequencing depth. If \code{size} is
larger than population, the input \code{x} will be returned.}

\item{rep}{Repeat sampling \code{rep} times and take the average.}
}
\value{
A \code{vector}, \code{matrix}, or other object.
}
\description{
Random subsampling a matrix, or a given object with abundance or
count table. This command is a generic function that treat different input
differently. Read \strong{Details} for more information.
}
\details{
This function performs differently based on different input. If a
\code{character vector} is given, it will random pick \code{size} elements
without replacement. If a \code{integer vector} is given, the function will
first create a vector for population pool based on the \code{names} and
\code{count} and then random pick \code{size} elements without replacement.
If a \code{double vector} is given, the function will do sampling WITH
replacement by using relative abundance as probabilities. The two different
implementations for \code{integer} and \code{double} are generally the same
when \code{rep} is sufficiently large. But when both population size and
time of \code{rep} are extremely small, \code{double} method may generate
odd result. Other inputs will utilize above three methods.
}
\note{
Some objects like \code{\link[Matrix]{Matrix}} will automatically
transfer \code{integer} to \code{double}. So shuffle the matrix first before
assign it to object who use \code{Matrix} in \code{physet} if you really
feel uncomfortable with using probability for \code{double}. Note that
shuffle only 1 time DO increase bias while shuffled data may cause bias for
richness estimation. To solve these problem, \code{physet} will normalize
the otu_table with seqdep for most of the statistical analysis and will use
raw otu_table for richness rarefy analysis. Also as shuffle large \code{rep}
times is extremly time consuming for high-dimensional data and the result
is just as same as normalizing the data by sequencing depth. So it is not
recommended to use this method unless it is really necessary.
}
\examples{
data(oral)
sub <- .subsampling(oral, 100, 100)
sub <- .subsampling(as.matrix(otu_table(oral)), 100, 10)
}
\references{
phyloseq-paper
}
\seealso{
\code{\link{sample}}, \code{\link[permute]{shuffle}}
}
\keyword{internal}

