% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/physet-class.R
\docType{data}
\name{phylo-class}
\alias{phylo}
\alias{phylo-class}
\title{Inherited and Virtual Classes}
\description{
Define virtual classes and inherit \code{ape::phylo-class}
}
\details{
S3 Class \code{\link[ape]{phylo}} is not formally defined in ape
package. In order to inherit all the methods in \code{\link{ape}} package,
an empty list is registered as \dQuote{phylo} and extended as a S4 virtual
class allow \code{NULL}.
}
\seealso{
\code{\link[ape]{phylo}}, \code{\link{setOldClass}}
}
\keyword{internal}

